Add download middleware
